backend/requirements.txt
djangorestframework
django-cors-headers
matplotlib
torch
transformers
scikit-learn
pandas
opencv-python
deepface
moviepy
SpeechRecognition
tf-keras
